Move display-test boostrap
