+++ deobsf/v26_1/src/forge/java/org/adde0109/pcf/v26_1/forge/DisplayTestBootstrap.java
@@ -1,4 +1,4 @@
-package org.adde0109.pcf.v17_1.forge;
+package org.adde0109.pcf.v26_1.forge;
 
 import net.minecraftforge.fml.IExtensionPoint;
 
@@ -7,12 +7,13 @@ import org.jspecify.annotations.NonNull;
 import java.util.function.BiPredicate;
 import java.util.function.Supplier;
 
-public final class PCFBootstrap {
+public final class DisplayTestBootstrap {
     public static Supplier<@NonNull String> IGNORE_SERVER_ONLY =
             () -> {
                 throw new IllegalStateException("DisplayTest value Not initialized");
             };
 
+    @SuppressWarnings("unused")
     public static final BiPredicate<@NonNull String, Boolean> REMOTE_VERSION_TEST =
             (remoteVersion, isFromServer) -> true;
