+++ modern/v17_1/build.gradle.kts
@@ -20,6 +20,7 @@ unimined.minecraft(forge) {
 
 dependencies {
     forgeCompileOnly(srcSetAsDep(":modern:v16_5", "forge"))
+    forgeCompileOnly(srcSetAsDep(":deobsf:v26_1", "forge"))
     forgeCompileOnly(project(":common"))
 }
